fix: resolve MSVC C4244 and C4267 warnings on Windows
C4244 (simd_math_helpers.h:222): _mm512_kand returns __mmask16 regardless
of argument types; _kand_mask8 is the correct 8-bit mask AND intrinsic and
produces __mmask8 directly, eliminating the narrowing.

C4267 (two_state_hmm.h, test_fb_mode_parity.cpp): DiscreteDistribution
constructor takes int; callers were passing constexpr std::size_t values.
Add static_cast<int> at both sites. The int constructor is retained to
preserve negative-argument validation (DiscreteDistribution(-1) must throw).

Verified: 0 warnings, 46/46 tests pass on Windows/MSVC/AVX-512.

+++ examples/support/two_state_hmm.h
@@ -45,13 +45,15 @@ inline void prepare_two_state_hmm(Hmm &hmm) {
     hmm.setPi(pi);
     hmm.setTrans(trans);
 
-    auto fairDist = std::make_unique<DiscreteDistribution>(dishonest_casino::num_symbols);
+    auto fairDist =
+        std::make_unique<DiscreteDistribution>(static_cast<int>(dishonest_casino::num_symbols));
     for (std::size_t i = 0; i < dishonest_casino::num_symbols; ++i) {
         fairDist->setProbability(static_cast<int>(i), dishonest_casino::fair_emission_prob);
     }
     hmm.setDistribution(dishonest_casino::fair_state, std::move(fairDist));
 
-    auto loadedDist = std::make_unique<DiscreteDistribution>(dishonest_casino::num_symbols);
+    auto loadedDist =
+        std::make_unique<DiscreteDistribution>(static_cast<int>(dishonest_casino::num_symbols));
     for (std::size_t i = 0; i < dishonest_casino::num_symbols - 1; ++i) {
         loadedDist->setProbability(static_cast<int>(i), dishonest_casino::loaded_normal_prob);
     }
